pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "armodel"
dynamic = ["version"]
description = "the python arxml parser"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "melodypapa", email = "melodypapa@outlook.com"},
]
keywords = ["AUTOSAR", "ARXML"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "colorama",
    "openpyxl",
    "lxml",
]

[project.optional-dependencies]
pytest = ["pytest-cov"]

[project.urls]
Homepage = "http://github.com/melodypapa/py-armodel"

[project.scripts]
format-xml = "armodel.cli.format_xml_cli:main"
arxml-dump = "armodel.cli.arxml_dump_cli:cli_main"
arxml-format = "armodel.cli.arxml_format_cli:main"
connector2xlsx = "armodel.cli.connector2xlsx_cli:main"
connector-update = "armodel.cli.connector_update_cli:main"
armodel-component = "armodel.cli.swc_list_cli:main"
armodel-system-signal = "armodel.cli.system_signal_cli:main"
armodel-memory-section = "armodel.cli.memory_section_cli:main"
armodel-file-list = "armodel.cli.file_list_cli:main"
armodel-uuid-checker = "armodel.cli.uuid_checker_cli:main"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "armodel.__version__"}

[tool.ruff]
# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

# Line length to match coding standards (79 chars per PEP 8)
# CI warns at 127 but standard is 79 per CODING_RULE_LAYOUT_00002
line-length = 79

# Target Python version (ruff minimum is py37, project requires >=3.9)
target-version = "py37"

[tool.ruff.lint]
# Enable rules based on coding standards
# E: pycodestyle errors
# W: pycodestyle warnings
# F: Pyflakes
# I: isort (import sorting)
# N: pep8-naming (except N802 for AUTOSAR camelCase methods)
# B: flake8-bugbear
# C4: flake8-comprehensions
# SIM: flake8-simplify
select = ["E", "W", "F", "I", "N", "B", "C4", "SIM"]

# Ignore wildcard import violations for M2 model structure
# This is intentional per project coding rules (CODING_RULE_STYLE_00008)
# Also ignore AUTOSAR camelCase methods (N802) - they follow AUTOSAR standard
# Also ignore module naming (N999) - AUTOSAR modules use PascalCase
# Also ignore pyupgrade (UP) - project requires Python 3.5+ compatibility
# Also ignore line length issues (E501, W505) - project allows longer lines for clarity
ignore = [
    "N802",  # AUTOSAR methods use camelCase (not snake_case)
    "N999",  # AUTOSAR modules use PascalCase (not lowercase_with_underscores)
    "E501",  # Line too long - project allows longer lines for clarity
    "W505",  # Doc line too long - project allows longer docstrings for clarity
]

# Configure specific rule settings
[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files (wildcard export pattern)
# "__init__.py" = ["F401"]

# V2 models: Ignore specific rules due to AUTOSAR patterns and M2 structure
# E501/W505: Line/docstring too long - unavoidable due to deep M2 package structure
# E402: Module imports not at top - intentional to avoid circular dependencies in class definitions
# E721: Type comparisons with == - intentional pattern for abstract class checks
# SIM102: Collapsible if - style preference for nested if statements
# B024: Abstract class without abstract methods - some AUTOSAR classes are abstract by convention
# B904: Raise without from - minor exception handling style
# F811: Redefined while unused - intentional class redefinitions
# N803/N815: Naming - AUTOSAR camelCase conventions differ from PEP 8
#"src/armodel/models_v2/**/*.py" = [
#    "E501", "W505", "E402", "E721", "SIM102",
#    "B024", "B904", "F811", "N803", "N815"
#]

# V2 models (src/armodel/v2): Specific ignores for AUTOSAR patterns
# E402: Module imports not at top - intentional to avoid circular dependencies in class definitions
# E721: Type comparisons with == - intentional pattern for abstract class checks
# SIM102: Collapsible if - style preference for nested if statements
# B024: Abstract class without abstract methods - some AUTOSAR classes are abstract by convention
# B904: Raise without from - minor exception handling style
# F811: Redefined while unused - intentional class redefinitions
# N803/N815: Naming - AUTOSAR camelCase conventions differ from PEP 8
"src/armodel/v2/**/*.py" = [
    "E402", "E721", "SIM102",
    "B024", "B904", "F811", "N803", "N815"
]

[tool.ruff.lint.isort]
# Import order: Standard library → Third-party → Local (CODING_RULE_IMPORT_00001)
known-first-party = ["armodel"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pycodestyle]
# Max docstring length: 72 characters (CODING_RULE_LAYOUT_00002)
max-doc-length = 72

[tool.mypy]
# MyPy configuration for V2 models type checking
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter checking in future
disallow_incomplete_defs = false  # Set to true for stricter checking in future
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"  # or "skip" for stub-less packages
ignore_missing_imports = false

# Exclude V2 modules from type checking
exclude = [
    "src/armodel/v2/.*",
]

# Per-module mypy configuration
[[tool.mypy.overrides]]
module = [
    "armodel.v2.models.*",
]
disallow_untyped_defs = true  # V2 models should have type hints but allow gradual adoption
check_untyped_defs = true

# Ignore specific modules or directories
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "scripts.*",
]
ignore_errors = false  # Set to true to skip type checking in tests temporarily
